\name{print.pairwise.htest}
\alias{print.pairwise.htest}

\title{
Print pairwise.htest object in HTML format
}
\description{
Over-rides the print.pairwise.htest command to print pairwise.htest objects in HTML
}
\usage{
\method{print}{pairwise.htest}(x, ...)
}
\arguments{
  \item{x}{
A \code{pairwise.htest} object
}
  \item{\dots}{
Other commands passed to print command.
}
}

\value{
A character vector of length 1 containing the code to print the results of a pairwise.htest in HTML.
}
\author{
Bill Altermatt
}
